+++ public/comment.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
     <link rel="stylesheet" href="./css/comment.css" />
-    <script defer src="./js/comment.js"></script>
+    <script type="module" src="./comment/app.mjs"></script>
   </head>
   <body>
     <article class="box-comment">
