+++ public/utils/date.js
@@ -1,5 +1,7 @@
 export const todayFormat = () => {
-  const today = new Date();
   const format = num => (num < 10 ? `0${num}` : `${num}`);
-  return `${today.getFullYear()}/${format(today.getMonth() + 1)}/${format(today.getDate())}`;
+
+  const today = new Date();
+  // Todo: Time
+  return `${today.getFullYear()}-${format(today.getMonth() + 1)}-${format(today.getDate())}`;
 };
